+++ src/utilities.jl
@@ -13,3 +13,23 @@ function mean_abs_dev{T}(a::Array{T,1}, scale::Bool=false)
 
     mean(res) * c
 end
+
+"""
+Parameters:
+    * orig: the original colnames
+    * suffix: the suffix applied to `orig`
+
+```jldoctest
+julia> gen_colnames(["Open", "Close"], ["macd", "dif", "sig"])
+6-element Array{String,1}:
+ "Open_macd"
+ "Close_macd"
+ "Open_dif"
+ "Close_dif"
+ "Open_sig"
+ "Close_sig"
+```
+"""
+function gen_colnames(orig::Vector{String}, suffix::Vector{String})
+   vec(["$o\_$s" for o ∈ orig, s ∈ suffix])
+end